+++ src/black/concurrency.py
@@ -78,6 +78,7 @@ def reformat_many(
     mode: Mode,
     report: Report,
     workers: Optional[int],
+    no_cache: bool = False,
 ) -> None:
     """Reformat multiple files using a ProcessPoolExecutor."""
     maybe_install_uvloop()
@@ -115,6 +116,7 @@ def reformat_many(
                 report=report,
                 loop=loop,
                 executor=executor,
+                no_cache=no_cache,
             )
         )
     finally:
@@ -134,6 +136,7 @@ async def schedule_formatting(
     report: "Report",
     loop: asyncio.AbstractEventLoop,
     executor: "Executor",
+    no_cache: bool = False,
 ) -> None:
     """Run formatting of `sources` in parallel using the provided `executor`.
 
@@ -142,9 +145,11 @@ async def schedule_formatting(
     `write_back`, `fast`, and `mode` options are passed to
     :func:`format_file_in_place`.
     """
-    cache = Cache.read(mode)
-
-    if write_back in (WriteBack.DIFF, WriteBack.COLOR_DIFF):
+    cache = None if no_cache else Cache.read(mode)
+    if cache is None or write_back in (
+        WriteBack.DIFF,
+        WriteBack.COLOR_DIFF,
+    ):
         sources_with_stats: Union[set[Path], set[tuple[Optional[os.stat_result], Path]]] = sources
     else:
         sources_with_stats, cached = cache.filtered_cached(sources)
@@ -198,7 +203,7 @@ async def schedule_formatting(
                 report.done(src, changed)
     if cancelled:
         await asyncio.gather(*cancelled, return_exceptions=True)
-    if sources_to_cache:
+    if sources_to_cache and not no_cache and cache is not None:
         cache.write(sources_to_cache)